# docs/fullstack/api-contract/mobile/openapi.yaml
openapi: 3.1.0
info:
  title: LifeLink KH — Mobile API
  version: 0.3.0
  description: |
    Flutter donor/requester app. Machine twin of contract.md — this file wins on conflict.
    Scope: the 8 core FRs only (docs/scope.md). Deferred FRs have no endpoints here.

    Three response rules enforced across every schema below:
      1. No donor-facing response contains latitude or longitude (ADR 0003).
      2. distanceKm is server-rounded to 0.5.
      3. Contact details appear only after a donor has ACCEPTED (TM-AUTH-001 I1).
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /matches/me:
    get:
      tags: [match]
      summary: Donor inbox — requests this donor is matched to
      responses:
        '200':
          description: Matches
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Match' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /matches/{id}/respond:
    post:
      tags: [match]
      summary: Accept or decline
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [response]
              properties:
                response:
                  type: string
                  enum: [ACCEPTED, DECLINED]
      responses:
        '200':
          description: |
            Recorded. requesterContact is returned only for ACCEPTED, and phoneVerified is
            always false in this build (ADR 0002).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MatchResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Not this donor's match }
        '409': { description: Already responded }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    Unauthorized:
      description: Missing, invalid, expired, or foreign-project token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    BloodType:
      type: string
      description: The 8 ABO/Rh values. No "unknown" — it has no blood_compatibility row.
      enum: ['O-', 'O+', 'A-', 'A+', 'B-', 'B+', 'AB-', 'AB+']
...
    Urgency:
      type: string
      enum: [CRITICAL, URGENT, ROUTINE]
    RequestStatus:
      type: string
      description: EXPIRED is unreachable in this build — FR-REQUEST-005 deferred (DEC-004).
      enum: [OPEN, FULFILLED, CANCELLED]
    DistanceKm:
      type: number
      description: Server-rounded to 0.5 km (ADR 0003). Null when the donor has no coordinates.
      multipleOf: 0.5
      nullable: true
      example: 2.5
...
    DistrictName:
      type: object
      description: >
        Both labels, not one. The app switches language at runtime (FR-GLOBAL-001), so a server
        that picked a language would force a re-fetch on every locale change.
      required: [km, en]
      properties:
        km: { type: string, example: ទួលគោក }
        en: { type: string, example: Tuol Kouk }
...
    Hospital:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Calmette Hospital }
        districtName:
          allOf:
            - $ref: '#/components/schemas/DistrictName'
          nullable: true
          description: Both labels (CR-MAPI-003). Null until the hospital seed assigns a district.
...
    BloodRequest:
      type: object
      required: [id, status, patientBloodType, unitsNeeded, urgency, hospital, alertedCount, acceptedCount, createdAt]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/RequestStatus' }
        patientBloodType: { $ref: '#/components/schemas/BloodType' }
        unitsNeeded: { type: integer }
        urgency: { $ref: '#/components/schemas/Urgency' }
        hospital: { $ref: '#/components/schemas/Hospital' }
        alertedCount:
          type: integer
          description: request_matches rows written. The number the waiting screen shows.
        acceptedCount: { type: integer }
        createdAt: { type: string, format: date-time }
    RequesterContact:
      type: object
      description: |
        Returned only to a donor whose match response is ACCEPTED (TM-AUTH-001 I1).
        Never present before acceptance, on any endpoint.
      required: [displayName, phone, phoneVerified]
      properties:
        displayName: { type: string }
        phone: { type: string }
        phoneVerified:
          type: boolean
          description: |
            Always false in this build. Phone numbers stopped being verified when auth moved
            to Google Sign-In (ADR 0002); the client must show the caveat.
    BloodRequestDetail:
      allOf:
        - $ref: '#/components/schemas/BloodRequest'
        - type: object
          properties:
            # districtName was removed here in CR-MAPI-003 — hospital.districtName already
            # carries it, and two fields holding the same value drift.
            distanceKm: { $ref: '#/components/schemas/DistanceKm' }
            requesterContact:
              allOf:
                - $ref: '#/components/schemas/RequesterContact'
              nullable: true
    Match:
      type: object
      required: [matchId, request, myBloodType]
      properties:
        matchId: { type: string, format: uuid }
        request: { $ref: '#/components/schemas/BloodRequestDetail' }
        myBloodType:
          $ref: '#/components/schemas/BloodType'
        response:
          type: string
          enum: [ACCEPTED, DECLINED]
          nullable: true
        notifiedAt: { type: string, format: date-time, nullable: true }
    MatchResponse:
      type: object
      required: [matchId, response, respondedAt]
      properties:
        matchId: { type: string, format: uuid }
        response: { type: string, enum: [ACCEPTED, DECLINED] }
        respondedAt: { type: string, format: date-time }
        requesterContact:
          allOf:
            - $ref: '#/components/schemas/RequesterContact'
          nullable: true
...
    Error:
      type: object
      description: |
        Messages never contain a phone number, blood type, token, or coordinate
        (TM-AUTH-001 I2).
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: ROLE_NOT_ALLOWED }
            message: { type: string, example: Role must be DONOR or REQUESTER }
security:
  - bearerAuth: []